TMP: Move from float to double

+++ cuda_lin_alg/src/cuda_utils.cu
@@ -2,15 +2,15 @@
 
 #include <cuda_runtime.h>
 
-float* transfer_to_cuda(const float* host_array, const size_t count) {
-    float* device_memory;
+double* transfer_to_cuda(const double* host_array, const size_t count) {
+    double* device_memory;
     cudaMalloc(&device_memory, count);
     cudaMemcpy(device_memory, host_array, count, cudaMemcpyHostToDevice);
     return device_memory;
 }
 
-float* transfer_from_cuda(const float* device_array, const size_t count) {
-    auto* host_memory = static_cast<float*>(malloc(sizeof(float) * count));
+double* transfer_from_cuda(const double* device_array, const size_t count) {
+    auto* host_memory = static_cast<double*>(malloc(sizeof(double) * count));
     cudaMemcpy(host_memory, device_array, count, cudaMemcpyDeviceToHost);
     return host_memory;
 }
