+++ cuda_lin_alg/src/gpu_matrix.h
@@ -3,20 +3,20 @@
 #include <cuda_runtime.h>
 
 extern "C" {
-void launch_tiled_multiply(const float* A,
+void launch_tiled_multiply(const double* A,
         const unsigned int ai,
         const unsigned int aj,
-        const float* B,
+        const double* B,
         const unsigned int bj,
-        float* C,
+        double* C,
         const dim3& grid,
         const dim3& block,
         const unsigned int shared_mem_size);
 }
 
-__global__ void tiled_multiply(const float* A,
+__global__ void tiled_multiply(const double* A,
         const unsigned int ai,
         const unsigned int aj,
-        const float* B,
+        const double* B,
         const unsigned int bj,
-        float* C);
+        double* C);
