+++ lin_alg/src/matrix.cpp
@@ -18,12 +18,12 @@ bool Dimension::operator!=(const Dimension& other) const {
     return !(*this == other);
 }
 
-Matrix::Matrix(float* impl, const Dimension& dim)
+Matrix::Matrix(double* impl, const Dimension& dim)
     : data_{impl}
     , dim_{dim} {
 }
 
-Matrix Matrix::from_raw(float* impl, const Dimension& dim) {
+Matrix Matrix::from_raw(double* impl, const Dimension& dim) {
     return Matrix{impl, dim};
 }
 
@@ -32,22 +32,22 @@ Matrix::~Matrix() {
 }
 
 Matrix Matrix::zeroes(const Dimension& dim) {
-    return Matrix{static_cast<float*>(calloc(dim.i * dim.j, sizeof(float))), dim};
+    return Matrix{static_cast<double*>(calloc(dim.i * dim.j, sizeof(double))), dim};
 }
 
-Matrix Matrix::all_same(float entry, const Dimension& dim) {
-    auto* entries = static_cast<float*>(malloc(dim.i * dim.j * sizeof(float)));
+Matrix Matrix::all_same(double entry, const Dimension& dim) {
+    auto* entries = static_cast<double*>(malloc(dim.i * dim.j * sizeof(double)));
     for (auto index = 0U; index < dim.i * dim.j; ++index) {
         entries[index] = entry;
     }
     return Matrix{entries, dim};
 }
 
-float Matrix::operator()(const unsigned int i, const unsigned int j) const {
+double Matrix::operator()(const unsigned int i, const unsigned int j) const {
     return data_[i * dim_.j + j];
 }
 
-float& Matrix::operator()(const unsigned int i, const unsigned int j) {
+double& Matrix::operator()(const unsigned int i, const unsigned int j) {
     return data_[i * dim_.j + j];
 }
 
@@ -81,7 +81,7 @@ bool Matrix::operator==(const Matrix& other) const {
     return true;
 }
 
-const float* Matrix::raw() const {
+const double* Matrix::raw() const {
     return data_;
 }
